Moved generated artifact cleanup to ArtifactManager
