+++ crates/sui-replay-2/src/artifacts/manager.rs
@@ -2,13 +2,16 @@
 // SPDX-License-Identifier: Apache-2.0
 
 use crate::artifacts::{MoveCallInfo, ReplayCacheSummary};
-use anyhow::{Result, anyhow, bail};
+use anyhow::{Context, Result, anyhow, bail};
 use move_trace_format::format::{MoveTrace, MoveTraceReader};
 use std::{
+    fs,
     io::Write,
     path::{Path, PathBuf},
 };
-use sui_types::{effects::TransactionEffects, gas::GasUsageReport};
+use sui_types::{base_types::ObjectID, effects::TransactionEffects, gas::GasUsageReport};
+
+pub(crate) const BCODE_DIR: &str = "bytecode";
 
 pub const ARTIFACTS_ENCODING_EXT: &str = "json";
 pub const ARTIFACTS_ENCODING_COMPRESSION_EXT: &str = "json.zst";
@@ -49,7 +52,8 @@ pub struct ArtifactManager<'a> {
     pub base_path: &'a Path,
     pub overrides_allowed: bool,
     /// When true, all write operations (`serialize_artifact`, `serialize_move_trace`,
-    /// `try_remove_artifact`) become no-ops and the base directory is not created.
+    /// `try_remove_artifact`, `clear_generated_artifacts`) become no-ops and the base directory is
+    /// not created.
     /// Reads still work if the base directory already exists. Used for performance
     /// profiling where artifact serialization would otherwise dominate the trace.
     pub skip_writes: bool,
@@ -107,6 +111,86 @@ impl<'b> ArtifactManager<'b> {
             manager: self,
             artifact_type: artifact,
         }
+    }
+
+    /// Remove generated artifacts and package bytecode while preserving source and unknown files.
+    pub(crate) fn clear_generated_artifacts(&self) -> Result<()> {
+        if self.skip_writes || !self.overrides_allowed {
+            return Ok(());
+        }
+
+        let recovery = format!(
+            "move or remove the replay artifact directory at {} manually, then retry",
+            self.base_path.display()
+        );
+        let entries = match fs::read_dir(self.base_path) {
+            Ok(entries) => entries,
+            Err(error) if error.kind() == std::io::ErrorKind::NotFound => return Ok(()),
+            Err(error) => {
+                return Err(error).with_context(|| {
+                    format!(
+                        "failed to inspect prior artifacts at {}; {recovery}",
+                        self.base_path.display()
+                    )
+                });
+            }
+        };
+
+        for entry in entries {
+            let entry = entry.with_context(|| {
+                format!(
+                    "failed to inspect prior artifacts at {}; {recovery}",
+                    self.base_path.display()
+                )
+            })?;
+            let file_type = entry.file_type().with_context(|| {
+                format!(
+                    "failed to inspect prior artifact at {}; {recovery}",
+                    entry.path().display()
+                )
+            })?;
+            let file_name = entry.file_name();
+            if !file_type.is_dir()
+                || file_name
+                    .to_str()
+                    .and_then(|name| name.parse::<ObjectID>().ok())
+                    .is_none()
+            {
+                continue;
+            }
+
+            let bytecode_path = entry.path().join(BCODE_DIR);
+            match fs::remove_dir_all(&bytecode_path) {
+                Ok(()) => {}
+                Err(error) if error.kind() == std::io::ErrorKind::NotFound => {}
+                Err(error) => {
+                    return Err(error).with_context(|| {
+                        format!(
+                            "failed to clear prior generated bytecode at {}; {recovery}",
+                            bytecode_path.display()
+                        )
+                    });
+                }
+            }
+        }
+
+        for artifact in ARTIFACTS {
+            let artifact_path = self.base_path.join(artifact.as_file());
+            match fs::remove_file(&artifact_path) {
+                Ok(()) => {}
+                Err(error) if error.kind() == std::io::ErrorKind::NotFound => {}
+                Err(error) => {
+                    return Err(error).with_context(|| {
+                        format!(
+                            "failed to clear prior artifact at {}; {recovery}",
+                            artifact_path.display()
+                        )
+                    });
+                }
+            }
+        }
+
+        Ok(())
     }
 }
 
@@ -351,5 +435,52 @@ impl ArtifactMember<'_, '_> {
             );
         }
         Ok(())
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    /// Cleanup removes generated artifacts and bytecode while preserving source and other user
+    /// files. A missing digest directory is also accepted.
+    #[test]
+    fn clears_generated_artifacts_and_preserves_user_files() {
+        let temp = tempfile::tempdir().unwrap();
+        let artifact_path = temp.path().join("digest");
+        let artifact_manager = ArtifactManager::new(&artifact_path, true).unwrap();
+        let package_path = artifact_path.join(format!("{:?}", ObjectID::random()));
+        let bytecode_path = package_path.join(BCODE_DIR);
+        let source_path = package_path.join("source");
+        let user_path = artifact_path.join("notes");
+        fs::create_dir_all(&bytecode_path).unwrap();
+        fs::create_dir_all(&source_path).unwrap();
+        fs::create_dir_all(user_path.join(BCODE_DIR)).unwrap();
+        fs::write(bytecode_path.join("stale.mv"), b"stale").unwrap();
+        fs::write(source_path.join("package.move"), b"source").unwrap();
+        fs::write(user_path.join(BCODE_DIR).join("notes.txt"), b"notes").unwrap();
+        for artifact in ARTIFACTS {
+            fs::write(artifact_path.join(artifact.as_file()), b"stale").unwrap();
+        }
+
+        artifact_manager.clear_generated_artifacts().unwrap();
+
+        assert!(artifact_path.exists());
+        assert!(!bytecode_path.exists());
+        assert_eq!(
+            fs::read(source_path.join("package.move")).unwrap(),
+            b"source"
+        );
+        assert_eq!(
+            fs::read(user_path.join(BCODE_DIR).join("notes.txt")).unwrap(),
+            b"notes"
+        );
+        for artifact in ARTIFACTS {
+            assert!(!artifact_path.join(artifact.as_file()).exists());
+        }
+
+        artifact_manager.clear_generated_artifacts().unwrap();
+        fs::remove_dir_all(&artifact_path).unwrap();
+        artifact_manager.clear_generated_artifacts().unwrap();
     }
 }
