+++ crates/sui-replay-2/src/simulation.rs
@@ -14,15 +14,11 @@
 //! reported as a warning rather than presenting the local replay as authoritative.
 
 use crate::{
-    artifacts::{ARTIFACTS, ArtifactManager},
+    artifacts::ArtifactManager,
     replay_txn::{ExecutorProvider, replay_transaction_from_store},
-    tracing::BCODE_DIR,
 };
-use anyhow::{Context, Error, Result, ensure};
-use std::{
-    fs,
-    path::{Path, PathBuf},
-};
+use anyhow::{Error, Result, ensure};
+use std::path::{Path, PathBuf};
 use sui_data_store::{
     EpochData, EpochStore, Node, ObjectKey, ObjectStore, TransactionInfo, TransactionStore,
     VersionQuery, stores::DataStore,
@@ -109,9 +105,10 @@ pub fn trace_simulated_transaction(
     let digest = transaction.digest();
     let network = node.network_name();
     let artifact_path = output_root.join(digest.to_string());
+    let artifact_manager = ArtifactManager::new(&artifact_path, true)?;
 
     // A rerun must not expose generated artifacts left by an earlier attempt with the same digest.
-    clear_generated_artifacts(&artifact_path)?;
+    artifact_manager.clear_generated_artifacts()?;
 
     // This checkpoint bounds only reads for which transaction execution provides no version.
     let source = DataStore::new(node.clone(), user_agent)?;
@@ -131,7 +128,6 @@ pub fn trace_simulated_transaction(
         mock_gas_object,
     };
 
-    let artifact_manager = ArtifactManager::new(&artifact_path, true)?;
     let mut executor_provider = ExecutorProvider::new(false);
     // Reuse historical replay with tracing enabled and its standard artifact behavior.
     let outcome = replay_transaction_from_store(
@@ -151,81 +147,6 @@ pub fn trace_simulated_transaction(
         trace_generated: outcome.trace_generated,
         mock_gas_id,
     })
-}
-
-fn clear_generated_artifacts(path: &Path) -> Result<()> {
-    let recovery = format!(
-        "move or remove the replay artifact directory at {} manually, then retry",
-        path.display()
-    );
-    let entries = match fs::read_dir(path) {
-        Ok(entries) => entries,
-        Err(error) if error.kind() == std::io::ErrorKind::NotFound => return Ok(()),
-        Err(error) => {
-            return Err(error).with_context(|| {
-                format!(
-                    "failed to inspect prior artifacts at {}; {recovery}",
-                    path.display()
-                )
-            });
-        }
-    };
-
-    for entry in entries {
-        let entry = entry.with_context(|| {
-            format!(
-                "failed to inspect prior artifacts at {}; {recovery}",
-                path.display()
-            )
-        })?;
-        let file_type = entry.file_type().with_context(|| {
-            format!(
-                "failed to inspect prior artifact at {}; {recovery}",
-                entry.path().display()
-            )
-        })?;
-        let file_name = entry.file_name();
-        if !file_type.is_dir()
-            || file_name
-                .to_str()
-                .and_then(|name| name.parse::<ObjectID>().ok())
-                .is_none()
-        {
-            continue;
-        }
-
-        let bytecode_path = entry.path().join(BCODE_DIR);
-        match fs::remove_dir_all(&bytecode_path) {
-            Ok(()) => {}
-            Err(error) if error.kind() == std::io::ErrorKind::NotFound => {}
-            Err(error) => {
-                return Err(error).with_context(|| {
-                    format!(
-                        "failed to clear prior generated bytecode at {}; {recovery}",
-                        bytecode_path.display()
-                    )
-                });
-            }
-        }
-    }
-
-    for artifact in ARTIFACTS {
-        let artifact_path = path.join(artifact.as_file());
-        match fs::remove_file(&artifact_path) {
-            Ok(()) => {}
-            Err(error) if error.kind() == std::io::ErrorKind::NotFound => {}
-            Err(error) => {
-                return Err(error).with_context(|| {
-                    format!(
-                        "failed to clear prior artifact at {}; {recovery}",
-                        artifact_path.display()
-                    )
-                });
-            }
-        }
-    }
-
-    Ok(())
 }
 
 impl TransactionStore for SimulationReplayStore {
@@ -508,45 +429,5 @@ mod tests {
         assert_eq!(local_mock_gas(Some(&object), &wrong_id), None);
         assert_eq!(local_mock_gas(Some(&object), &checkpoint), None);
         assert_eq!(local_mock_gas(None, &exact), None);
-    }
-
-    /// Cleanup removes generated artifacts and bytecode while preserving source and other user
-    /// files. A missing digest directory is also accepted.
-    #[test]
-    fn clears_generated_artifacts_and_preserves_user_files() {
-        let temp = tempfile::tempdir().unwrap();
-        let artifact_path = temp.path().join("digest");
-        let package_path = artifact_path.join(format!("{:?}", ObjectID::random()));
-        let bytecode_path = package_path.join(BCODE_DIR);
-        let source_path = package_path.join("source");
-        let user_path = artifact_path.join("notes");
-        std::fs::create_dir_all(&bytecode_path).unwrap();
-        std::fs::create_dir_all(&source_path).unwrap();
-        std::fs::create_dir_all(user_path.join(BCODE_DIR)).unwrap();
-        std::fs::write(bytecode_path.join("stale.mv"), b"stale").unwrap();
-        std::fs::write(source_path.join("package.move"), b"source").unwrap();
-        std::fs::write(user_path.join(BCODE_DIR).join("notes.txt"), b"notes").unwrap();
-        for artifact in ARTIFACTS {
-            std::fs::write(artifact_path.join(artifact.as_file()), b"stale").unwrap();
-        }
-
-        clear_generated_artifacts(&artifact_path).unwrap();
-
-        assert!(artifact_path.exists());
-        assert!(!bytecode_path.exists());
-        assert_eq!(
-            std::fs::read(source_path.join("package.move")).unwrap(),
-            b"source"
-        );
-        assert_eq!(
-            std::fs::read(user_path.join(BCODE_DIR).join("notes.txt")).unwrap(),
-            b"notes"
-        );
-        for artifact in ARTIFACTS {
-            assert!(!artifact_path.join(artifact.as_file()).exists());
-        }
-
-        clear_generated_artifacts(&artifact_path).unwrap();
-        clear_generated_artifacts(&temp.path().join("missing")).unwrap();
     }
 }
